Migrate common utilities to new format.
