+++ front/lib/actions/constants.ts
@@ -27,9 +27,6 @@ export const DEFAULT_CONVERSATION_SEARCH_ACTION_NAME =
 
 export const DEFAULT_PROJECT_SEARCH_ACTION_NAME = "search_project_context";
 
-export const SEARCH_AVAILABLE_USERS_TOOL_NAME = "search_available_users";
-export const GET_MENTION_MARKDOWN_TOOL_NAME = "get_mention_markdown";
-
 export const DUST_CONVERSATION_HISTORY_MAGIC_INPUT_KEY =
   "__dust_conversation_history";
